Add files via upload
This file contains a task management tool. the tool can be used to sort tasks for a user according to their preferences and priorities. To use the tool, launch the html page. A prompt window appears with a dialogue box for the task title and the priority level. After imputing the task and priority, click on the add tasks box to add the task. After completing the task, click on the task to strikethrough and show that it's completed.