Create BMI Calculator
This folder contains a body/ mass index calculator.
To use the tool, run the html page.  
Dialogue boxes appear asking for the height and mass information from the user.
After imputing the information, click on the calculate tab to calculate the bmi.
The bmi index appears with the accompanying health status according to the information.